feat: Setting Initial FileStructure
- 기능 구현에 필요한 기본 파일 구조를 설정하였음.
