+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/Portfolio.java
@@ -4,13 +4,14 @@ import com.kakao.sunsuwedding.user.planner.Planner;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
-@Getter
-@NamedEntityGraph(name = "PortfolioWithPlanner",
-                  attributeNodes = @NamedAttributeNode("planner"))
 @Entity
+@Getter
+@NoArgsConstructor
+@NamedEntityGraph(name = "PortfolioWithPlanner", attributeNodes = @NamedAttributeNode("planner"))
 public class Portfolio {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -67,9 +68,6 @@ public class Portfolio {
         this.minPrice = minPrice;
         this.maxPrice = maxPrice;
         this.createdAt = (createdAt == null? LocalDateTime.now() : createdAt);
-    }
-
-    protected Portfolio() {
     }
 
     public void updateTitle(String title) {
