+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/PortfolioRequest.java
@@ -1,0 +1,4 @@
+package com.kakao.sunsuwedding.portfolio;
+
+public class PortfolioRequest {
+}
