+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/price/PriceItemJPARepository.java
@@ -1,0 +1,6 @@
+package com.kakao.sunsuwedding.portfolio.price;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PriceItemJPARepository extends JpaRepository <PriceItem, Integer> {
+}
